+++ albums/photoslibrary_repository.go
@@ -85,10 +85,14 @@ func (r PhotosLibraryAlbumsRepository) Get(ctx context.Context, albumId string) 
 	return &album, nil
 }
 
+// maxItemsPerPage is the maximum number of albums to ask to the PhotosLibrary. Fewer albums might
+// be returned than the specified number. See https://developers.google.com/photos/library/guides/list#pagination
+const maxItemsPerPage = 50
+
 // ListAll fetches and caches all the albums from the repo.
 func (r PhotosLibraryAlbumsRepository) ListAll(ctx context.Context) ([]Album, error) {
 	albumsResult := make([]Album, 0)
-	err := r.service.List().ExcludeNonAppCreatedData().Pages(ctx, func(response *photoslibrary.ListAlbumsResponse) error {
+	err := r.service.List().ExcludeNonAppCreatedData().PageSize(maxItemsPerPage).Pages(ctx, func(response *photoslibrary.ListAlbumsResponse) error {
 		for _, res := range response.Albums {
 			albumsResult = append(albumsResult, toAlbum(res))
 		}